Setting buffer_sizes to ull

+++ comms/src/blazingdb/transport/io/reader_writer.cpp
@@ -89,7 +89,7 @@ void setPinnedBufferProvider(std::size_t sizeBuffers, std::size_t numBuffers) {
 PinnedBufferProvider &getPinnedBufferProvider() { return *global_instance; }
 
 void writeBuffersFromGPUTCP(std::vector<ColumnTransport> &column_transport,
-                            std::vector<int> bufferSizes,
+                            std::vector<unsigned long long> bufferSizes,
                             std::vector<const char *> buffers, void *fileDescriptor,
                             int gpuNum) {
   if (bufferSizes.size() == 0) {
